% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2.R
\name{FindMaxDiff}
\alias{FindMaxDiff}
\title{Calculate maxdiff  nMaxdiff}
\usage{
FindMaxDiff(data, control, original, rounded, datareturn = FALSE)
}
\arguments{
\item{data}{data}

\item{control}{control}

\item{original}{original}

\item{rounded}{rounded}

\item{datareturn}{datareturn}
}
\value{
maxdiff and nMaxdiff
}
\description{
Calculate maxdiff  nMaxdiff
}
\keyword{internal}
