% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSrounding.R
\name{FormulaSelection.PLSrounded}
\alias{FormulaSelection.PLSrounded}
\title{FormulaSelection  method for PLSrounded}
\usage{
\method{FormulaSelection}{PLSrounded}(x, formula = NULL, intercept = NA, logical = FALSE)
}
\arguments{
\item{x}{PLSrounded object}

\item{formula}{\code{formula} parameter to \code{\link[SSBtools]{FormulaSelection}}.
When \code{NULL} (default), the publish data frame is returned without any limitation.}

\item{intercept}{\code{intercept} parameter to \code{FormulaSelection}.}

\item{logical}{\code{logical} parameter to \code{FormulaSelection}.}
}
\value{
Limited version of the publish data frame
}
\description{
FormulaSelection  method for PLSrounded
}
