% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2.R
\name{MakeControl}
\alias{MakeControl}
\title{Make suggested control - input to makeroundtabs()}
\usage{
MakeControl(d, data, level = 2)
}
\arguments{
\item{d}{as input to \code{makeroundtabs()}}

\item{data}{data.frame}

\item{level}{Interaction level}
}
\value{
list
}
\description{
Make suggested control - input to makeroundtabs()
}
\keyword{internal}
