% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noeFraKostraFiler.R
\encoding{UTF8}
\name{MakeHierFormula}
\alias{MakeHierFormula}
\title{Make model formula from data taking into account hierarchical variables}
\usage{
MakeHierFormula(data = NULL, hGroups = HierarchicalGroups2(data),
  n = length(hGroups), sim = TRUE)
}
\arguments{
\item{data}{data frame}

\item{hGroups}{Output from HierarchicalGroups2()}

\item{n}{Interaction level or 0 (all levels)}

\item{sim}{Include "~" when TRUE}
}
\value{
Formula as character string
}
\description{
Make model formula from data taking into account hierarchical variables
}
\examples{
 z3 <- SmallCountData("z3")
 MakeHierFormula(z3[,-7])
 MakeHierFormula(z3[,-7],n=2)
 MakeHierFormula(z3[,-7],n=0)
}
