% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlsRound.R
\name{PlsRoundSparse}
\alias{PlsRoundSparse}
\title{PlsRoundSparse}
\usage{
PlsRoundSparse(x, roundBase = 3, yInner, yPublish = Matrix::crossprod(x,
  yInner)[, 1, drop = TRUE], singleRandom = FALSE, maxIter = 1e+06,
  maxIterRows = 1000)
}
\arguments{
\item{x}{Model matrix}

\item{roundBase}{roundBase}

\item{yInner}{inner cells}

\item{yPublish}{cells to be published}

\item{singleRandom}{Single random draw when TRUE}

\item{yInnerExact}{Original yInner (when iteration)}

\item{yPublishExact}{Original yPublish (when iteration)}
}
\value{
rounded versions of yInner and yPublish
}
\description{
Avrunder basert på en algoritme inspirert av PLS-regresjon som forutsetter dummy matrise (Model matrix)
}
\keyword{internal}
