% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSrounding.R
\name{print.PLSrounded}
\alias{print.PLSrounded}
\title{Print method for PLSrounded}
\usage{
\method{print}{PLSrounded}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{PLSrounded object}

\item{digits}{positive integer.  Minimum number of significant digits to be used for printing most numbers.}

\item{\dots}{further arguments sent to the underlying}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for PLSrounded
}
\keyword{print}
