% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2.R
\name{RoundViaDummy2}
\alias{RoundViaDummy2}
\title{RoundViaDummy2}
\usage{
RoundViaDummy2(
  data,
  b,
  d,
  nin,
  micro = FALSE,
  control = NULL,
  allTerms = FALSE,
  singleRandom = FALSE
)
}
\arguments{
\item{data}{data}

\item{b}{b}

\item{d}{d}

\item{nin}{nin}

\item{micro}{micro}

\item{control}{control}

\item{allTerms}{Use all interaction terms in formula instead of using control}

\item{singleRandom}{Single random draw when TRUE (instead of algorithm)}
}
\value{
Output from RoundViaDummy extended with  "yControl" "maxdiff"  "nMaxdiff" "formula"
}
\description{
RoundViaDummy with input as makeroundtabs
}
\keyword{internal}
