% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseModelMatrix.R
\name{SparseModelMatrix}
\alias{SparseModelMatrix}
\title{Denne er erstattet av FormulaSums/ModelMatrix Overparameterized sparse model matrix with possible cross table}
\usage{
SparseModelMatrix(formula, data = NULL, mf = model.frame(formula, data =
  data), hg = HierarchicalGroups3(mf), makeColnames = TRUE,
  crossTable = FALSE, total = "Total", sep = "-", printInc = TRUE)
}
\arguments{
\item{formula}{A model formula}

\item{data}{data frame}

\item{mf}{model frame (alternative input instead of data)}

\item{hg}{List defining hierarchical variable groups}

\item{makeColnames}{Column names made when TRUE}

\item{crossTable}{Cross table in output when TRUE}

\item{total}{String used to name totals}

\item{sep}{String to separate when creating column names}

\item{printInc}{Printing "..." to console when TRUE}
}
\value{
The model matrix or a list with model matrix and cross table.
}
\description{
All factor levels included. Constructed by calling fac2sparse() repeatedly instead of sparse.model.matrix().
Hierarchical variables handled when constructing cross table. Column names constructed from the cross table.
}
\examples{
  z1 <- SmallCountData("z1")
  ModelMatrix(~region*hovedint,z1)
}
\keyword{internal}
