% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2.R
\name{Lists2formula}
\alias{Lists2formula}
\title{Make formula from input parameters to makeroundtabs()}
\usage{
Lists2formula(d, control = NULL, data = NULL)
}
\arguments{
\item{d}{d}

\item{control}{control}

\item{data}{data}
}
\value{
formula as string
}
\description{
Make formula from input parameters to makeroundtabs()
}
\keyword{internal}
