% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSroundingLoop.R
\name{PLSroundingLoop}
\alias{PLSroundingLoop}
\title{PLSrounding on portions of data at a time}
\usage{
PLSroundingLoop(
  data,
  loopId,
  ...,
  zeroCandidates = FALSE,
  preRounded = NULL,
  plsWeights = NULL,
  printInc = TRUE,
  preDifference = TRUE
)
}
\arguments{
\item{data}{data}

\item{loopId}{loopId}

\item{...}{dots}

\item{zeroCandidates}{zeroCandidates}

\item{preRounded}{preRounded}

\item{plsWeights}{plsWeights}

\item{printInc}{printInc}

\item{preDifference}{preDifference}
}
\value{
As output from \code{\link{PLSrounding}}
}
\description{
PLSrounding on portions of data at a time
}
\examples{
mf2 <- ~region + fylke * hovedint
z2 <- SmallCountData("z2")
a <- PLSroundingLoop(z2, loopId = "kostragr", freqVar = "ant", formula = mf2)
a
}
