% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallCountData.R
\encoding{UTF8}
\name{SmallCountData}
\alias{SmallCountData}
\title{Function that returns a dataset}
\usage{
SmallCountData(dataset, path = NULL)
}
\arguments{
\item{dataset}{Name of data set within the SmallCountRounding  package}

\item{path}{When non-NULL the data set is read from "path/dataset.RData"}
}
\value{
The dataset
}
\description{
Function that returns a dataset
}
\note{
Currently, the function returns the same datasets as is included in the package easySdcTable.
}
\examples{
 z  <- SmallCountData("sosialFiktiv")

}
